+++ apps/web/src/app/components/ControlsBar.tsx
@@ -32,7 +32,7 @@ import { memo, useCallback, useEffect, useRef, useState, type FormEvent } from "
 import {
   BACKGROUND_EFFECT_OPTIONS,
   type BackgroundEffect,
-} from "../lib/background-blur";
+} from "../lib/camera-effects";
 import type {
   MeetingConfigSnapshot,
   MeetingUpdateRequest,
