+++ apps/web/src/app/components/mobile/MobileJoinScreen.tsx
@@ -17,7 +17,7 @@ import {
   createManagedCameraTrack,
   type BackgroundEffect,
   type ManagedCameraTrack,
-} from "../../lib/background-blur";
+} from "../../lib/camera-effects";
 import type { ConnectionState, MeetError } from "../../lib/types";
 import { DEFAULT_AUDIO_CONSTRAINTS } from "../../lib/constants";
 import {
