+++ apps/web/src/app/lib/camera-effects/media.ts
@@ -1,0 +1,51 @@
+"use client";
+
+import {
+  LOW_QUALITY_CONSTRAINTS,
+  STANDARD_QUALITY_CONSTRAINTS,
+} from "../constants";
+import type { VideoQuality } from "../types";
+
+export const getVideoConstraints = (
+  quality: VideoQuality,
+): MediaTrackConstraints => {
+  return quality === "low"
+    ? { ...LOW_QUALITY_CONSTRAINTS }
+    : { ...STANDARD_QUALITY_CONSTRAINTS };
+};
+
+export const stopMediaStream = (stream: MediaStream) => {
+  stream.getTracks().forEach((track) => {
+    track.onended = null;
+    try {
+      track.stop();
+    } catch {}
+  });
+};
+
+export const waitForVideoReady = async (
+  video: HTMLVideoElement,
+): Promise<void> => {
+  if (video.readyState >= HTMLMediaElement.HAVE_CURRENT_DATA) {
+    return;
+  }
+
+  await new Promise<void>((resolve, reject) => {
+    const handleLoadedData = () => {
+      cleanup();
+      resolve();
+    };
+    const handleError = () => {
+      cleanup();
+      reject(new Error("Camera preview failed to start"));
+    };
+    const cleanup = () => {
+      video.removeEventListener("loadeddata", handleLoadedData);
+      video.removeEventListener("error", handleError);
+    };
+
+    video.addEventListener("loadeddata", handleLoadedData, { once: true });
+    video.addEventListener("error", handleError, { once: true });
+  });
+};
+
